requirements.txt
pandas==1.1.3
numpy
Streamlit 
plotly
folium
matplotlib
mapclassify
osmnx
streamlit_folium
openpyxl
sklearn
google-api-python-client
google-auth
google-auth-httplib2
protobuf
click==8